+++ src/main/java/com/themore/moamoatown/job/dto/JobRequestsResponseDTO.java
@@ -1,4 +1,4 @@
-package com.themore.moamoatown.town.dto;
+package com.themore.moamoatown.job.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
