+++ src/main/java/com/themore/moamoatown/town/controller/TownController.java
@@ -2,7 +2,10 @@ package com.themore.moamoatown.town.controller;
 
 import com.themore.moamoatown.common.annotation.MemberId;
 import com.themore.moamoatown.common.annotation.TownId;
-import com.themore.moamoatown.town.dto.*;
+import com.themore.moamoatown.town.dto.TownCreateInternalDTO;
+import com.themore.moamoatown.town.dto.TownCreateRequestDTO;
+import com.themore.moamoatown.town.dto.TownCreateResponseDTO;
+import com.themore.moamoatown.town.dto.TownTaxResponseDTO;
 import com.themore.moamoatown.town.service.TownService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j;
@@ -10,7 +13,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
-import java.util.List;
 
 /**
  * 타운 컨트롤러
@@ -25,7 +27,6 @@ import java.util.List;
  * 2024.08.23   임원정        타운 만들기 추가
  * 2024.08.24   임원정        타운 만들기 메소드 수정
  * 2024.08.26   임원정        타운 세금 현황 조회 추가
- * 2024.08.26   임원정        타운 역할 신청 현황 조회 추가
  * </pre>
  */
 
@@ -69,17 +70,6 @@ public class TownController {
     @GetMapping("/tax")
     public ResponseEntity<TownTaxResponseDTO> getTotalTax(@TownId Long townId) {
         TownTaxResponseDTO response = townService.getTotalTax(townId);
-        return ResponseEntity.ok(response);
-    }
-
-    /**
-     * 타운 역할 신쳥 현황 조회
-     * @param townId
-     * @return
-     */
-    @GetMapping("/job/request")
-    public ResponseEntity<List<JobRequestsResponseDTO>> getJobRequests(@TownId Long townId) {
-        List<JobRequestsResponseDTO> response = townService.getJobRequests(townId);
         return ResponseEntity.ok(response);
     }
 }
