+++ src/main/java/com/themore/moamoatown/town/service/TownService.java
@@ -1,11 +1,8 @@
 package com.themore.moamoatown.town.service;
 
-import com.themore.moamoatown.town.dto.JobRequestsResponseDTO;
 import com.themore.moamoatown.town.dto.TownCreateInternalDTO;
 import com.themore.moamoatown.town.dto.TownCreateRequestDTO;
 import com.themore.moamoatown.town.dto.TownTaxResponseDTO;
-
-import java.util.List;
 
 /**
  * 타운 서비스 인터페이스
@@ -19,12 +16,10 @@ import java.util.List;
  * 2024.08.23  	임원정        최초 생성
  * 2024.08.23   임원정        createTown 메소드 추가
  * 2024.08.26   임원정        getTotalTax 메소드 추가
- * 2024.08.26   임원정        getJobRequests 메소드 추가
  * </pre>
  */
 
 public interface TownService {
     TownCreateInternalDTO createTown(TownCreateRequestDTO townCreateRequestDTO, Long memberId);
     TownTaxResponseDTO getTotalTax(Long townId);
-    List<JobRequestsResponseDTO> getJobRequests(Long townId);
 }
